refactor(web): move top bar components outside of layout folder
